Style fix
